backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-generativeai>=0.5.4
requests>=2.31.0
langgraph>=0.2.0
scikit-learn>=1.4.0
pandas==2.1.4
joblib==1.3.2
python-dotenv==1.0.0
pydantic>=2.5.3,<2.13
pydantic-settings==2.1.0
email-validator==2.1.0
statsmodels>=0.14.0

